pyproject.toml
[project]
name = "websearxng-tool"
version = "0.1.0"
description = "A flexible MCP server and library for SearXNG and web scraping."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ddgs>=9.5.2",
    "fastmcp>=2.14.2",
    "mcp>=1.25.0",
    "pdfminer-six>=20240706",
    "playwright>=1.54.0",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "rich>=14.2.0",
    "trafilatura>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
websearx = "websearx_tool.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/websearx_tool"]
